+++ examples/7_multi_pane/multi_pane.py
@@ -11,7 +11,8 @@ def calculate_sma(df, period: int = 50, name = None):
 
 
 def demo():
-    # chart = HTMLChart(width=1200, height=800, inner_height=-500, filename='charts.html')
+    # chart = HTMLChart(width=1200, height=800, inner_height=-500)
+    # chart.export('charts.html')
     chart = Chart(width=1200, height=800, title='Multi Pane Demo')
     chart.legend(visible=True)
     df = pd.read_csv('ohlcv.csv')